Refactor sudo checking
